+++ src/queries/useAuthorities/useAuthorities.test.js
@@ -142,7 +142,7 @@ describe('Given useAuthorities', () => {
 
       await waitFor(() => !result.current.isLoading);
 
-      expect(result.current.query).toEqual('(identifiers.value=="n  00000001 ")');
+      expect(result.current.query).toEqual('(identifiers.value=="n  00000001 " and authRefType=="Authorized")');
     });
   });
 
@@ -166,7 +166,7 @@ describe('Given useAuthorities', () => {
         sortedColumn: '',
       }), { wrapper });
 
-      expect(result.current.query).toEqual('(keyword=="advancedTest1") not (identifiers.value=="advancedTest2")');
+      expect(result.current.query).toEqual('(keyword=="advancedTest1") not (identifiers.value=="advancedTest2" and authRefType=="Authorized")');
     });
   });
 });